// frontend/inventoryTracking/src/app/Models/product.ts
export class Product {
    product_id: String;
    product_name: String;
    product_type: String;
    product_brand: String;
    product_lable: String;
    product_start_inv: String;
    product_in_inv: String;
    product_min_req: String;
    product_unit_pr: String;
    product_sell_pr:String;
}